requirements.txt
pandas
tqdm
Pillow
python-dotenv
google-generativeai
ipython
ipykernel
ipywidgets